+++ src/common/Random.hpp
@@ -10,8 +10,8 @@ class Random {
   public:
   explicit Random();
 
-  size_t
-  Range(size_t min, size_t max);
+  uint32_t
+  Range(uint32_t min, uint32_t max);
 
   template <typename T>
   const auto&
@@ -19,16 +19,16 @@ class Random {
 
   private:
   std::mt19937_64 Rng;
-  std::uniform_int_distribution<size_t> dist;
+  std::uniform_int_distribution<uint32_t> dist;
 };
 
 inline Random::Random() {
   Rng.seed(static_cast<uint64_t>(std::chrono::steady_clock::now().time_since_epoch().count()));
 }
 
-inline size_t
-Random::Range(size_t min, size_t max) {
-  dist.param(std::uniform_int_distribution<size_t>::param_type(min, max));
+inline uint32_t
+Random::Range(uint32_t min, uint32_t max) {
+  dist.param(std::uniform_int_distribution<uint32_t>::param_type(min, max));
   return dist(Rng);
 }
